Shortened location names.
Before all the location names were fully qualified with their parent locations (e.g. Hospital Name Ward Named Bed 01). After a browse through the UI we realised that there is always going to be the parent name there to provide context as this is a field on the model and the desktop is just views of these models. It therefore seemed worth it to short these names and reduce redundant wording in the UI to clean things up a bit.