Locked down patient model permissions.
A patient model should only be created by a registration. It therefore makes sense to give user groups access to registration but not the patient model itself. The implementation can then use the super user for the creation but nothing else, this keeps behaviour constrained and makes exploits less likely.

Decided to take the concept behaviour by removing everything except read permissions on the patient model from all user groups. Any interaction with the patient model is likely to be done through another model (such as registration) so there should be no need for these permissions.