Removed `.gitattributes`.
We're not sure why the Git attributes file was addded but it specifies
that line feed is always used as the end-of-line character. Local
setting until this file was checked out was for Git to automatically
handle converting LF to CRLF and vice versa. This meant that when the
new Git attributes file was detected Git modified and added a lot of
image files to the working tree that it thought had CRLF that needed
changing to honour the new setting. For some reason it was impossible to
get rid of these modified images from the working tree. Removing the
unneeded Git attributes file resolved this.