Fixed bug in refusal frequencies.
The refusal frequencies could only be one thing depending on the risk but in reality there can be multiple frequencies for a given risk other than the default frequency that is in the configuration (may need to clarify this with the name of the configuration parameter). Escalation tasks and policy customisation based on admission date are example of things that can change the frequency for a patient to something other than the default frequency for their risk.

Changing the logic to take the frequency of the refused observation instead means that although the risk is not known we are honouring any frequency changes for the patient and trusting that when they were changed that was correct.