Fixed test failing intermittently.
The logic for the expected time was incorrect. It added the frequency onto the create datetime of the new obs activity which occurs slightly later than the terminate datetime of the previous obs activity which is what should be used. This faulty logic sometimes works because the new observation activity is created very quickly but invariably the small amount of time it takes will sometimes take it across the threshold into a new second and cause the assertion afterwards to fail.