Attempt to use existing methods. Hasn't worked.
Figured that passing the IDs of all the patients on the ward to the existing `collect_patients` method might work but it did not. I believe this is because of the joins and `WHERE` clauses in the SQL query that is called. This means that I will have to re-implement the logic in the SQL in Python. This is something that needed doing anyway but will increase development time.