Observations can have a table but no graph.
Some logic already existed to accommodate this, for example the tabs would be hidden if only one of either the table or graph was defined for an observation. However one problem was that if only a table was defined then the 'Ranged Values' checkbox still showed and awkwardly overlapped the table. This option is only applicable to graphs and so it shouldn't have been displayed at all.